Disoplaying income based rejection letter
if income threshold exceeded
